+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Note that since we don't clearly distinguish between a public and private interf
 
 - Show histogram in direct volume control point settings
 - Add `solidInterior` parameter to sphere/cylinder impostors
-- Add `meshes` and `volseg` extensions
+- Add `meshes` and `volumes-and-segmentations` extensions
 
 ## [v3.27.0] - 2022-12-15